Replace the `checked_int_cast` dependency with `std::convert::TryInto`.
The `checked_int_cast` crate (version 1.0.0 released on 2015-09-07) has
been made obsolete with the introduction of the `std::convert::TryInfo`
trait (the RFC has been started on 2016-03-10 - see
https://rust-lang.github.io/rfcs/1542-try-from.html.  The `qrcode`
depends on Rust 2018 edition and therefore it can depend on the presence
of the `std::convert::TryInfo` trait.

Based on the above, this commit drops the `checked_int_cast` dependency
and replaces it with usage ot the standard `TryInto` trait.  This commit
should have no impact on the behavior, nor the API of the `qrcode`
crate.  OTOH, removing the dependency will have a small positive impact
on users of the `qrcode` crate - they no longer will have to fetch,
build, audit, etc. the `checked_int_cast` crate.